requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# OpenAI
openai==1.51.0

# Vector store — Qdrant Cloud
qdrant-client==1.11.3

# Redis Cloud
redis==5.1.1

# SQL — MySQL
sqlalchemy==2.0.35
pymysql==1.1.1

# PDF/Text extraction
unstructured==0.22.16
unstructured[pdf]==0.22.16
pdfminer.six==20251230

# Chunking / embedding utilities
numpy==2.1.3
tiktoken==0.8.0

# Config
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Windows — libmagic replacement (required by unstructured on Windows)
python-magic-bin==0.4.14